Extract Event\Meta and Venue\Meta classes from their Setup classes; both expose a single register() entry point with matching per-band protected helpers (#1532).
