+++ CLAUDE.md
@@ -112,7 +112,7 @@ GatherPress uses custom `post_type_supports` to decouple features from specific 
 
 **Venue post type supports** (declared on post types that act as venues):
 
-- `gatherpress-venue-information` — **Core venue identifier.** Registers five individual editor-writable post meta keys (`gatherpress_address`, `gatherpress_latitude`, `gatherpress_longitude`, `gatherpress_phone`, `gatherpress_website`), all `show_in_rest`, so venue fields can be bound to blocks via `core/post-meta` block bindings. Also registers eight server-populated structured-address meta keys (`gatherpress_house_number`, `gatherpress_street`, `gatherpress_city`, `gatherpress_county`, `gatherpress_state`, `gatherpress_postcode`, `gatherpress_country`, `gatherpress_country_code`) derived from `gatherpress_address` by an async geocode cron handler — these are `show_in_rest` for read access only; REST writes are silently stripped. The unprefixed field list is the `Venue\Setup::STRUCTURED_ADDRESS_FIELDS` constant — single source of truth for registration, REST stripping, the cron write loop, and `Venue::get_information()`. The cron is gated by `gatherpress_geocode_on_save_enabled` (opt-out) and `gatherpress_async_geocode_pre_enqueue_job` (Action Scheduler short-circuit). Also wires up the venue detail blocks and automatic `_gatherpress_venue` taxonomy term management. Meta revisions (`revisions_enabled`) are opted-in per post type and only applied when the post type declares `revisions` in its `supports` array. Declaring this support is what makes a post type a venue source.
+- `gatherpress-venue-information` — **Core venue identifier.** Registers five individual editor-writable post meta keys (`gatherpress_address`, `gatherpress_latitude`, `gatherpress_longitude`, `gatherpress_phone`, `gatherpress_website`), all `show_in_rest`, so venue fields can be bound to blocks via `core/post-meta` block bindings. Also registers eight server-populated structured-address meta keys (`gatherpress_house_number`, `gatherpress_street`, `gatherpress_city`, `gatherpress_county`, `gatherpress_state`, `gatherpress_postcode`, `gatherpress_country`, `gatherpress_country_code`) derived from `gatherpress_address` by an async geocode cron handler — these are `show_in_rest` for read access only; REST writes are silently stripped. The unprefixed field list is the `Venue\Meta::STRUCTURED_ADDRESS_FIELDS` constant — single source of truth for registration, REST stripping, the cron write loop, and `Venue::get_information()`. (The companion editor-writable list lives at `Venue\Meta::EDITOR_WRITABLE_FIELDS`.) Meta registration itself lives on `Venue\Meta::register()`, hooked on `registered_post_type` so it fires once per supported post type; `Event\Meta::register()` mirrors the shape for event-date and event-only meta. The cron is gated by `gatherpress_geocode_on_save_enabled` (opt-out) and `gatherpress_async_geocode_pre_enqueue_job` (Action Scheduler short-circuit). Also wires up the venue detail blocks and automatic `_gatherpress_venue` taxonomy term management. Meta revisions (`revisions_enabled`) are opted-in per post type and only applied when the post type declares `revisions` in its `supports` array. Declaring this support is what makes a post type a venue source.
 - `gatherpress-venue-map` — Map display meta (show/zoom/height) and the venue map block
 
 **How it works:**
