+++ includes/core/classes/venue/class-setup.php
@@ -20,13 +20,9 @@ use GatherPress\Core\Event\Event;
 use GatherPress\Core\Settings;
 use GatherPress\Core\Traits\Singleton;
 use GatherPress\Core\Utility;
-use GatherPress\Core\Validate;
-use GatherPress\Core\Venue\Map;
 use GatherPress\Core\Venue\Map\Setup as Map_Setup;
-use stdClass;
 use WP_Block_Patterns_Registry;
 use WP_Post;
-use WP_REST_Request;
 use WP_Term;
 
 /**
@@ -46,61 +42,6 @@ class Setup {
 	 * Enforces a single instance of this class.
 	 */
 	use Singleton;
-
-	/**
-	 * Suffixes of the editor-writable venue-information meta keys.
-	 *
-	 * These five keys are written by the editor (and by trusted server code
-	 * such as REST PATCH requests gated by `Utility::can_edit_post_meta()`).
-	 * Stored here as the unprefixed "field" form so consumers can either
-	 * iterate them directly (e.g. {@see Venue::get_information()}) or map
-	 * through {@see Utility::prefix_key()} to get the full meta keys.
-	 *
-	 * Single source of truth for: meta registration in
-	 * {@see self::maybe_register_post_meta()} and the editor-writable half
-	 * of `Venue::get_information()`. Pair with
-	 * {@see self::STRUCTURED_ADDRESS_FIELDS} for the readonly Photon-derived
-	 * counterpart; together the two arrays make up the full 13-field shape
-	 * returned by `Venue::get_information()`.
-	 *
-	 * @since 1.0.0
-	 * @var string[]
-	 */
-	public const EDITOR_WRITABLE_FIELDS = array(
-		'address',
-		'latitude',
-		'longitude',
-		'phone',
-		'website',
-	);
-
-	/**
-	 * Suffixes of the structured-address venue meta keys.
-	 *
-	 * These eight keys are derived from `gatherpress_address` by the async
-	 * geocode cron handler — populated server-side from the Photon response,
-	 * never written through the REST API. Stored here as the unprefixed
-	 * "field" form so consumers can either iterate them directly (e.g.
-	 * {@see Venue::get_information()}) or map through {@see Utility::prefix_key()}
-	 * to get the full meta keys (e.g. registration in
-	 * {@see self::maybe_register_post_meta()}).
-	 *
-	 * Single source of truth for: meta registration, REST readonly stripping,
-	 * cron handler write loop, and `Venue::get_information()` field list.
-	 *
-	 * @since 1.0.0
-	 * @var string[]
-	 */
-	public const STRUCTURED_ADDRESS_FIELDS = array(
-		'house_number',
-		'street',
-		'city',
-		'county',
-		'state',
-		'postcode',
-		'country',
-		'country_code',
-	);
 
 	/**
 	 * Class constructor.
@@ -130,6 +71,7 @@ class Setup {
 	 */
 	protected function instantiate_classes(): void {
 		Map_Setup::get_instance();
+		Meta::get_instance();
 	}
 
 	/**
@@ -148,7 +90,6 @@ class Setup {
 		);
 		add_action( 'init', array( $this, 'register_post_type' ) );
 		add_action( 'registered_post_type', array( $this, 'maybe_register_post_type_hooks' ) );
-		add_action( 'registered_post_type', array( $this, 'maybe_register_post_meta' ) );
 		// Priority 11 so post types registered at default priority 10 are available for get_post_types_by_support().
 		add_action( 'init', array( $this, 'register_taxonomy' ), 11 );
 		// post_updated has no per-type variant in WP core, and we need $post_before
@@ -303,232 +244,6 @@ class Setup {
 				),
 			)
 		);
-	}
-
-	/**
-	 * Sanitize callback for venue coordinate meta (latitude / longitude).
-	 *
-	 * Numeric values within the ±180 range are normalized to a string form via
-	 * a float cast (trims whitespace, normalizes scientific notation). Anything
-	 * else collapses to an empty string — the "no coords yet" sentinel the
-	 * editor and `Map::parse_coord()` already treat as unset.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param mixed $value The submitted value.
-	 * @return string Normalized coordinate string, or '' for invalid / unset input.
-	 */
-	public function sanitize_coordinate( $value ): string {
-		return Validate::coordinate( $value ) ? (string) (float) $value : '';
-	}
-
-	/**
-	 * Registers venue meta fields when a post type declares venue support.
-	 *
-	 * Meta is registered per support:
-	 * - gatherpress-venue-information: address, phone, website, lat/lng as
-	 *   individual keys so they can be bound directly via core/post-meta
-	 *   block bindings without an intermediate JSON parse step.
-	 * - gatherpress-venue-map: map display settings (show, zoom, height).
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param string $post_type The post type that was just registered.
-	 * @return void
-	 */
-	public function maybe_register_post_meta( string $post_type ): void {
-		// Structured-address pieces derived from `gatherpress_address` by an
-		// async cron handler that fires only when the address actually
-		// changes (via `updated_post_meta` short-circuit on no-op). These
-		// are server-populated from the geocoder, so REST writes are denied
-		// (`__return_false` auth_callback) and submitted values get stripped
-		// in `filter_readonly_meta` rather than triggering a permission
-		// error. Read access via REST stays open so JSON-LD / schema.org
-		// emitters and downstream API consumers can read them.
-		$structured_address_meta = array_map(
-			array( Utility::class, 'prefix_key' ),
-			self::STRUCTURED_ADDRESS_FIELDS
-		);
-
-		// Per-field sanitize_callback for the editor-writable venue-information
-		// fields. Keys must match `self::EDITOR_WRITABLE_FIELDS` exactly so
-		// the loop below can resolve a callback for each entry.
-		$editor_writable_sanitizers = array(
-			'address'   => 'sanitize_text_field',
-			'latitude'  => array( $this, 'sanitize_coordinate' ),
-			'longitude' => array( $this, 'sanitize_coordinate' ),
-			'phone'     => 'sanitize_text_field',
-			'website'   => 'sanitize_url',
-		);
-
-		// Map keeps a per-zoom descriptor map here: { "15": { url, hash },
-		// ... }. Exposed read-only via REST so the block editor can preview
-		// the cached static image when the user picks renderMode="static".
-		// Writes are denied — the server-side pipeline is the only thing
-		// allowed to populate this meta. Registered separately from the
-		// editor-writable loop because it's neither editor-writable nor
-		// share the standard string-meta args shape.
-		$map_meta_args = array(
-			'auth_callback' => '__return_false',
-			'show_in_rest'  => array(
-				'schema' => array(
-					'type'                 => 'object',
-					'additionalProperties' => array(
-						'type'       => 'object',
-						'properties' => array(
-							'url'  => array( 'type' => 'string' ),
-							'hash' => array( 'type' => 'string' ),
-						),
-					),
-				),
-			),
-			'single'        => true,
-			'type'          => 'object',
-		);
-
-		$venue_map_meta = array(
-			// Map display settings.
-			'gatherpress_map_show'   => array(
-				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
-				'sanitize_callback' => 'rest_sanitize_boolean',
-				'show_in_rest'      => true,
-				'single'            => true,
-				'type'              => 'boolean',
-				'default'           => true,
-			),
-			'gatherpress_map_zoom'   => array(
-				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
-				'sanitize_callback' => 'absint',
-				'show_in_rest'      => true,
-				'single'            => true,
-				'type'              => 'integer',
-				'default'           => 10,
-			),
-			'gatherpress_map_height' => array(
-				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
-				'sanitize_callback' => 'absint',
-				'show_in_rest'      => true,
-				'single'            => true,
-				'type'              => 'integer',
-				'default'           => 300,
-			),
-		);
-
-		if ( post_type_supports( $post_type, 'gatherpress-venue-information' ) ) {
-			$supports_revisions = post_type_supports( $post_type, 'revisions' );
-
-			// Editor-writable fields share a common args shape; only the
-			// per-field sanitize_callback varies. Iterate the constant so
-			// adding a new field is a single edit at the top of this class.
-			foreach ( self::EDITOR_WRITABLE_FIELDS as $field ) {
-				$args = array(
-					'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
-					'sanitize_callback' => $editor_writable_sanitizers[ $field ],
-					'show_in_rest'      => true,
-					'single'            => true,
-					'type'              => 'string',
-					'default'           => '',
-					'revisions_enabled' => true,
-				);
-
-				// revisions_enabled is only valid when the post type supports revisions.
-				// Silently drop it for venue post types that opt out (e.g. companion plugins
-				// registering a minimal venue post type without revisions support).
-				if ( ! $supports_revisions ) {
-					unset( $args['revisions_enabled'] );
-				}
-
-				register_post_meta( $post_type, Utility::prefix_key( $field ), $args );
-			}
-
-			// Map descriptors register on their own — different shape (object
-			// schema, no sanitize_callback) and not editor-writable.
-			register_post_meta( $post_type, Map::META_KEY, $map_meta_args );
-
-			// Structured-address meta share an identical args shape, so
-			// register them in a tight loop rather than duplicating the
-			// array literal eight times. `auth_callback` denies REST writes
-			// — these are populated by the async geocode cron handler, not
-			// the editor.
-			$structured_args = array(
-				'auth_callback'     => '__return_false',
-				'sanitize_callback' => 'sanitize_text_field',
-				'show_in_rest'      => true,
-				'single'            => true,
-				'type'              => 'string',
-				'default'           => '',
-				'revisions_enabled' => true,
-			);
-
-			if ( ! $supports_revisions ) {
-				unset( $structured_args['revisions_enabled'] );
-			}
-
-			foreach ( $structured_address_meta as $meta_key ) {
-				register_post_meta( $post_type, $meta_key, $structured_args );
-			}
-
-			// Strip read-only meta from REST requests so the editor can't write it directly.
-			// Belt-and-suspenders with `auth_callback => __return_false`: the
-			// strip filter runs in `rest_pre_insert_<post_type>` BEFORE the
-			// auth_callback would 403 the whole request, so a co-submitted
-			// PATCH with a writable field plus a readonly field succeeds for
-			// the writable subset rather than failing the whole payload.
-			// Both are required — removing either breaks the contract.
-			add_filter(
-				sprintf( 'rest_pre_insert_%s', $post_type ),
-				array( $this, 'filter_readonly_meta' ),
-				10,
-				2
-			);
-		}
-
-		if ( post_type_supports( $post_type, 'gatherpress-venue-map' ) ) {
-			foreach ( $venue_map_meta as $meta_key => $args ) {
-				register_post_meta( $post_type, $meta_key, $args );
-			}
-		}
-	}
-
-	/**
-	 * Filter out read-only meta from REST API requests.
-	 *
-	 * Some venue meta keys are populated by server-side pipelines (e.g. the
-	 * Map static map descriptors) and should never be written by the
-	 * block editor. Values submitted for those keys are silently discarded
-	 * rather than triggering a permission error from the __return_false auth
-	 * callback.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param stdClass        $prepared_post An object representing a single post prepared for inserting or updating.
-	 * @param WP_REST_Request $request       Request object.
-	 * @return stdClass The prepared post object.
-	 */
-	public function filter_readonly_meta( stdClass $prepared_post, WP_REST_Request $request ): stdClass {
-		// Structured-address fields are derived from `gatherpress_address`
-		// by the async geocode cron handler. REST writes are stripped
-		// rather than rejected so a client that PATCHes them alongside
-		// editor-writable fields doesn't fail the whole request.
-		$readonly_keys = array_merge(
-			array( Map::META_KEY ),
-			array_map(
-				array( Utility::class, 'prefix_key' ),
-				self::STRUCTURED_ADDRESS_FIELDS
-			)
-		);
-
-		$meta = $request->get_param( 'meta' );
-
-		if ( is_array( $meta ) ) {
-			foreach ( $readonly_keys as $key ) {
-				unset( $meta[ $key ] );
-			}
-
-			$request->set_param( 'meta', $meta );
-		}
-
-		return $prepared_post;
 	}
 
 	/**
